+++ src/models/database.py
@@ -1,35 +1,14 @@
 import sqlite3
 
-from src.common.filepaths import BLOGGING_DB
 from src.loggers.general_logger import GeneralLogger
 from src.common import filepaths
+from src.models.db_connection import DBConnection
 
-
-class DBConnection:
-    def __init__(self):
-        self.connection = None
-        self.host = BLOGGING_DB
-        self.cursor = None
-
-    def __enter__(self) -> sqlite3.Connection:
-        self.connection = sqlite3.connect(self.host)
-        self.cursor = self.connection.cursor()
-        return self.connection
-
-    def __exit__(self, exc_type, exc_val, exc_tb):
-        if exc_type or exc_val or exc_tb:
-            print(f"Exception type: {exc_type}")
-            print(f"Exception value: {exc_val}")
-            print(f"Exception traceback: {exc_tb}")
-            self.connection.close()
-
-        else:
-            self.connection.commit()
-            self.connection.close()
+blog_db_connection = DBConnection()
 
 
 def get_item(query, data):
-    with DBConnection() as cursor:
+    with blog_db_connection as cursor:
         try:
             response = cursor.execute(query, data).fetchone()
             return response
@@ -39,7 +18,7 @@ def get_item(query, data):
 
 
 def get_items(query, data=None):
-    with DBConnection() as cursor:
+    with blog_db_connection as cursor:
         try:
             if data is None:
                 response = cursor.execute(query).fetchall()
@@ -53,7 +32,7 @@ def get_items(query, data=None):
 
 
 def insert_item(query, data):
-    with DBConnection() as connection:
+    with blog_db_connection as connection:
         cursor = connection.cursor()
         try:
             cursor.execute(query, data)
